packages/dffmpeg-client/pyproject.toml
[project]
name = "dffmpeg-client"
version = "0.1.0"
authors = [
    { name="Bryan Wyatt", email="brwyatt@gmail.com" },
]
description = "Client application components for dffmpeg"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "dffmpeg-common",
    "httpx",
    "pyyaml",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["dffmpeg.*"]
namespaces = true

[project.optional-dependencies]
test = [
    "anyio",
    "black",
    "flake8",
    "isort",
    "pytest",
    "pytest-asyncio",
]
dev = [
    "dffmpeg-client[test]",
    "ipython",
]

[project.scripts]
dffmpeg-client = "dffmpeg.client.cli:main"
dffmpeg_proxy = "dffmpeg.client.cli:proxy_main"
